create module for indexes
