+++ fast_forward/__init__.py
@@ -207,4 +207,8 @@ python -m fast_forward.examples.interpolate \\
 ```
 """
 
+from fast_forward.index import Mode
+from fast_forward.index.in_memory_index import InMemoryIndex
+from fast_forward.ranking import Ranking
+
 __version__ = "0.1.0"
